Remove branch pruning
