+++ mythril/analysis/solver.py
@@ -1,10 +1,9 @@
 """This module contains analysis module helpers to solve path constraints."""
-from functools import lru_cache
 from typing import Dict, List, Tuple, Union
-from z3 import sat, unknown, FuncInterp
+from z3 import FuncInterp
 import z3
 
-from mythril.analysis.analysis_args import analysis_args
+from mythril.support.model import get_model
 from mythril.laser.ethereum.state.global_state import GlobalState
 from mythril.laser.ethereum.state.constraints import Constraints
 from mythril.laser.ethereum.keccak_function_manager import (
@@ -12,8 +11,7 @@ from mythril.laser.ethereum.keccak_function_manager import (
     hash_matcher,
 )
 from mythril.laser.ethereum.transaction import BaseTransaction
-from mythril.laser.smt import UGE, Optimize, symbol_factory
-from mythril.laser.ethereum.time_handler import time_handler
+from mythril.laser.smt import UGE, symbol_factory
 from mythril.exceptions import UnsatError
 from mythril.laser.ethereum.transaction.transaction_models import (
     ContractCreationTransaction,
@@ -21,44 +19,6 @@ from mythril.laser.ethereum.transaction.transaction_models import (
 import logging
 
 log = logging.getLogger(__name__)
-
-
-# LRU cache works great when used in powers of 2
-@lru_cache(maxsize=2 ** 23)
-def get_model(constraints, minimize=(), maximize=(), enforce_execution_time=True):
-    """
-
-    :param constraints:
-    :param minimize:
-    :param maximize:
-    :param enforce_execution_time: Bool variable which enforces --execution-timeout's time
-    :return:
-    """
-    s = Optimize()
-    timeout = analysis_args.solver_timeout
-    if enforce_execution_time:
-        timeout = min(timeout, time_handler.time_remaining() - 500)
-        if timeout <= 0:
-            raise UnsatError
-    s.set_timeout(timeout)
-    for constraint in constraints:
-        if type(constraint) == bool and not constraint:
-            raise UnsatError
-
-    constraints = [constraint for constraint in constraints if type(constraint) != bool]
-
-    for constraint in constraints:
-        s.add(constraint)
-    for e in minimize:
-        s.minimize(e)
-    for e in maximize:
-        s.maximize(e)
-    result = s.check()
-    if result == sat:
-        return s.model()
-    elif result == unknown:
-        log.debug("Timeout encountered while solving expression using z3")
-    raise UnsatError
 
 
 def pretty_print_model(model):
