+++ mythril/laser/ethereum/svm.py
@@ -203,6 +203,14 @@ class LaserEVM:
         self.time = datetime.now()
 
         for i in range(self.transaction_count):
+            if len(self.open_states) > 0:
+                old_states_count = len(self.open_states)
+                self.open_states = [
+                    state for state in self.open_states if state.constraints.is_possible
+                ]
+                prune_count = len(self.open_states) - old_states_count
+                if prune_count:
+                    log.info("Pruned {} states".format(prune_count))
             log.info(
                 "Starting message call transaction, iteration: {}, {} initial states".format(
                     i, len(self.open_states)
