+++ mythril/mythril/mythril_analyzer.py
@@ -9,9 +9,9 @@ from mythril.laser.ethereum.iprof import InstructionProfiler
 from . import MythrilDisassembler
 from mythril.support.source_support import Source
 from mythril.support.loader import DynLoader
+from mythril.support.support_args import args
 from mythril.analysis.symbolic import SymExecWrapper
 from mythril.analysis.callgraph import generate_graph
-from mythril.analysis.analysis_args import analysis_args
 from mythril.analysis.traceexplore import get_serializable_statespace
 from mythril.analysis.security import fire_lasers, retrieve_callback_issues
 from mythril.analysis.report import Report, Issue
@@ -67,8 +67,7 @@ class MythrilAnalyzer:
         self.enable_coverage_strategy = enable_coverage_strategy
         self.custom_modules_directory = custom_modules_directory
 
-        analysis_args.set_loop_bound(loop_bound)
-        analysis_args.set_solver_timeout(solver_timeout)
+        args.solver_timeout = solver_timeout
 
     def dump_statespace(self, contract: EVMContract = None) -> str:
         """
