+++ mythril/support/support_args.py
@@ -1,0 +1,13 @@
+from mythril.support.support_utils import Singleton
+
+
+class Args(object, metaclass=Singleton):
+    """
+    This module helps in preventing args being sent through multiple of classes to reach any analysis/z
+    """
+
+    def __init__(self):
+        self.solver_timeout = None
+
+
+args = Args()
